cleaned up format, got more compilation working, added chain weight code
