+++ src/algorithms/expected_consensus.md
@@ -48,8 +48,8 @@ In practice, EC defines two different fields within a block:
 - A `Ticket` field — this stores the new ticket generated during this block generation attempt. It is from this ticket that miners will sample randomness to run leader election in `K` rounds.
 - An `ElectionProof` — this stores a proof that a given miner has won a leader election using the appropriate ticket `K` rounds back appended with the current epoch number. It proves that the leader was validly elected in this epoch.
 
-{{< readfile file="expected_consensus.id" code="true" lang="go" >}}
-{{< readfile file="expected_consensus.go" code="true" lang="go" >}}
+{{< readfile file="election.id" code="true" lang="go" >}}
+{{< readfile file="election.go" code="true" lang="go" >}}
 
 ```
 But why the randomness lookback?
@@ -72,7 +72,11 @@ How is K selected?
 - On the other, making K smaller reduces adversarial power to grind.
 ```
 
+{{<label ticket_generation>}}
 ### Ticket generation
+
+{{< readfile file="expected_consensus.id" code="true" lang="go" >}}
+{{< readfile file="expected_consensus.go" code="true" lang="go" >}}
 
 This section discusses how tickets are generated by EC for the `Ticket` field.
 
@@ -97,7 +101,7 @@ Each Ticket should be generated from the prior one in the ticket-chain and verif
 {{<label leader_election>}}
 ## Secret Leader Election
 
-Expected Consensus is a consensus protocol that works by electing a miner from a weighted set in proportion to their power. In the case of Filecoin, participants and powers are drawn from the storage [power table](storage-market.md#the-power-table), where power is equivalent to storage provided through time.
+Expected Consensus is a consensus protocol that works by electing a miner from a weighted set in proportion to their power. In the case of Filecoin, participants and powers are drawn from the {{<sref power_table>}}, where power is equivalent to storage provided through time.
 
 Leader Election in Expected Consensus must be Secret, Fair and Verifiable. This is achieved through the use of randomness used to run the election. In the case of Filecoin's EC, the blockchain tracks an independent ticket chain. These tickets are used as randomness inputs for Leader Election. Every block generated references an `ElectionProof` derived from a past ticket. The ticket chain is extended by the miner who generates a new block for each successful leader election.
 
@@ -116,18 +120,25 @@ We use the VRF from {{<sref vrf>}} to run leader election in EC.
 
 If the miner wins the election in this round, it can use newEP, along with a newTicket to generate and publish a new block. Otherwise, it waits to hear of another block generated in this round.
 
-In short, the process of crafting a new ElectionProof in round N is as follows:
-TODO: ADD LINK
+In short, the process of crafting a new ElectionProof in round N is as follows in the `DrawElectionProof` function:
+{{< readfile file="storage_mining_subsystem.go" code="true" lang="go" >}}
 
 It is important to note that every block contains two artifacts: one, a ticket derived from last block's ticket to extend the ticket-chain, and two, an election proof derived from the ticket `K` rounds back used to run leader election.
 
 Note: We draw the miner power from the prior round. This means that a miner can win a block in an epoch in which they were supposed to have proven storage again but did not. Put another way, `ElectionProof` validity is checked prior to executing transactions (including that which cuts miner power) at every epoch.
 
-The miner can then check whether they drew a winning election proof by comparing their power fraction (taking into account the number of blocks to be crafted per epoch `e`) to the `ElectionProof`s value, as follows:
+The miner can then check whether they drew a winning election proof by comparing their power fraction to the `ElectionProof`s value, as follows:
 
-`ElectionProof/maxValue < e * minerPower/totalPower`
+`ElectionProof * TotalPower < e * MinerPower * MaxValue`
 
-using SHA-256 with our vdf, we would have `maxValue = 2^256`. The above is the integer format of `ElectionProof/2^256 < e * minerPower/totalPower` which makes it easy to see that a miner should win proportionally to the amount of power they have staked in the system. Note that miners with more than `1/e` fraction of power will win once at every round, meaning there may be fewer than e blocks per round with many large miners.
+with:
+- `ElectionProof`   the election proof's byte output
+- `TotalPower`      the total power in the current power table
+- `e`               the expected number of blocks crafted per epoch
+- `MinerPower`      the current miner's power in the power table
+- `MaxValue`        the maximum possible value for the election proof using SHA-256 with our vdf, we would have `maxValue = 2^256`.
+
+The above is the integer format of `ElectionProof/MaxValue < e * MinerPower/TotalPower` which makes it easy to see that a miner should win proportionally to the amount of power they have staked in the system. Note that miners with more than `1/e` fraction of power will win once at every round, meaning there may be fewer than e blocks per round with many large miners.
 
 If successful, the miner can craft a block, passing it to the block producer. If unsuccessful, it will wait to hear of another block mined this round to try again. In the case no other block was found in this round the miner can increment the epoch number and try leader election again using the same past ticket and new epoch number.
