+++ src/systems/filecoin_blockchain/storage_power_consensus/storage_power_consensus_subsystem.id
@@ -42,8 +42,8 @@ type StoragePowerConsensusSubsystem struct {//(@mutable)
     // call by StorageMiningSubsystem after sealing
     GetPoStChallenge(chain block.Chain, epoch block.ChainEpoch) base_mining.PoStChallenge
 
-    GetFinality()       block.ChainEpoch
-    FinalizedEpoch()    block.ChainEpoch
+    GetFinality()     block.ChainEpoch
+    FinalizedEpoch()  block.ChainEpoch
 }
 
 type StoragePowerConsensusError struct {}
